# packages/contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: Direkrut AI API
  version: 0.2.0
  description: |
    Dua layanan backend independen, didokumentasikan dalam satu file ini:

    - **apps/api-go** (`auth`, `jobs`, `subscriptions`, `payments`,
      `notifications`) -- API publik yang dipanggil langsung oleh
      `apps/web`. Endpoint listing/detail lowongan bersifat publik (gak
      butuh login); endpoint tulis butuh JWT.
    - **apps/ai-engine** (`cv-parser`, `assessment`, `vector-search`,
      `chat`) -- **internal-only**, dipanggil server-to-server oleh
      api-go, tidak pernah langsung dari browser. Trust boundary-nya
      bukan CORS, tapi shared-secret header `X-Internal-Api-Key`.

    Tiap operasi di bawah nunjukin server mana yang beneran ngelayanin
    dia lewat override `servers` per-path.

    ### Status implementasi
    Endpoint yang belum diimplementasi ditandai eksplisit di
    summary-nya (`[BELUM DIIMPLEMENTASI]`) dan balikin `501`. Semua yang
    lain di dokumen ini udah beneran jalan & udah dites end-to-end
    lawan database & object storage asli, bukan cuma spek di atas kertas.

    ### Bentuk error
    Kedua servis balikin bentuk error JSON yang SAMA di semua endpoint
    kecuali di-catat lain (lihat catatan tiap-endpoint buat pengecualian):
    ```json
    { "error": { "code": "validation_failed", "message": "..." } }
    ```

    ### Autentikasi
    - `bearerAuth` (JWT, di api-go): access token umur 15 menit. Klaim HRD
...
      kebocoran/replay).
...
      ai-engine.

...

...
servers:
...
paths:
...
  /v1/assessment/score-validation:
    servers:
      - url: https://direkrut-ai-ai-engine-c58d5404c946.herokuapp.com
        description: apps/ai-engine (production, Heroku)
      - url: http://localhost:8000
        description: apps/ai-engine (lokal)
    post:
      tags: [assessment]
      summary: Nilai jawaban validasi kompetensi kandidat
      description: >
        Skor 0-100 + breakdown deteksi keaslian jawaban (authentic vs
        generic vs aiGenerated, totalnya 100). Kalau model AI gak nurut
        format JSON, balikin nilai netral (0/0/0/0) bukan 500 -- tetap
        butuh review manual di kasus itu. Di-cache berdasar hash
        `application_id`+isi jawaban.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [application_id, responses]
              properties:
                application_id: { type: string }
                responses:
                  type: array
                  items:
                    type: object
                    required: [question, answer]
                    properties:
                      question: { type: string }
                      answer: { type: string }
      responses:
        "200":
          description: Skor & deteksi keaslian
          content:
            application/json:
              schema:
                type: object
                properties:
                  recommendation_score: { type: number, minimum: 0, maximum: 100 }
                  authenticity_score:
                    type: object
                    description: Totalnya 100.
                    properties:
                      authentic: { type: number }
                      generic: { type: number }
                      aiGenerated: { type: number }
        "401":
          $ref: "#/components/responses/InternalKeyInvalid"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/assessment/transcribe-interview:
    servers:
      - url: https://direkrut-ai-ai-engine-c58d5404c946.herokuapp.com
        description: apps/ai-engine (production, Heroku)
      - url: http://localhost:8000
        description: apps/ai-engine (lokal)
    post:
      tags: [assessment]
      summary: Speech-to-text + ringkasan analisis wawancara
      description: >
        Audio di-download dari object storage, ditranskrip lewat Groq
        Whisper, lalu diringkas 2-3 kalimat. Di-cache berdasar hash
        `application_id`+`audio_object_key`.
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [application_id, audio_object_key]
              properties:
                application_id: { type: string }
                audio_object_key: { type: string }
      responses:
        "200":
          description: Transkrip + ringkasan
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcript: { type: string }
                  analysis_summary: { type: string }
        "401":
          $ref: "#/components/responses/InternalKeyInvalid"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dipakai di apps/api-go. Access token umur 15 menit dari /v1/auth/login atau /v1/auth/register.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
      description: Dipakai di apps/ai-engine. Shared secret server-to-server, bukan buat dipanggil browser.
  responses:
...
    RateLimited:
      description: Kelebihan request dalam window waktu (lihat info.description)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalKeyInvalid:
      description: Header X-Internal-Api-Key gak ada / salah
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: validation_failed }
            message: { type: string, example: "request body gak valid" }
...